requirements.txt
requests==2.24.0
flask==1.1.2
flask==0.14.3
psycopg2-binary==2.8.5
flask-migrate==2.5.3
flask-sqlalchemy==2.4.4
# requests